Move app logic out of lib
