+++ src/app.rs
@@ -1,0 +1,252 @@
+#[cfg(target_arch = "wasm32")]
+use std::{cell::RefCell, rc::Rc};
+#[cfg(target_arch = "wasm32")]
+use wasm_bindgen::{prelude::*, closure::Closure, JsCast};
+#[cfg(target_arch = "wasm32")]
+use wgpu::util::DeviceExt;
+use glam::Mat4;
+
+#[cfg(target_arch = "wasm32")]
+use crate::input::camera::Camera;
+use crate::graphics::{self, vertex::{VERTICES, INDICES, as_bytes}};
+
+#[cfg(target_arch = "wasm32")]
+#[repr(C)]
+#[derive(Clone, Copy)]
+struct Uniforms {
+    mvp: [[f32; 4]; 4],
+}
+
+#[cfg(target_arch = "wasm32")]
+struct State {
+    surface: wgpu::Surface<'static>,
+    device: wgpu::Device,
+    queue: wgpu::Queue,
+    pipeline: wgpu::RenderPipeline,
+    vertex_buffer: wgpu::Buffer,
+    index_buffer: wgpu::Buffer,
+    uniform_buffer: wgpu::Buffer,
+    bind_group: wgpu::BindGroup,
+    depth_texture: wgpu::Texture,
+    depth_view: wgpu::TextureView,
+    depth_format: wgpu::TextureFormat,
+    aspect: f32,
+}
+
+#[cfg(target_arch = "wasm32")]
+impl State {
+    async fn new(canvas: &web_sys::HtmlCanvasElement) -> Result<Self, JsValue> {
+        let instance = wgpu::Instance::default();
+        let surface = instance
+            .create_surface(wgpu::SurfaceTarget::Canvas(canvas.clone()))
+            .map_err(|e| JsValue::from_str(&e.to_string()))?;
+        let adapter = instance
+            .request_adapter(&wgpu::RequestAdapterOptions {
+                power_preference: wgpu::PowerPreference::HighPerformance,
+                compatible_surface: Some(&surface),
+                force_fallback_adapter: false,
+            })
+            .await
+            .ok_or("failed to find adapter")?;
+        let (device, queue) = adapter
+            .request_device(
+                &wgpu::DeviceDescriptor {
+                    label: None,
+                    required_features: wgpu::Features::empty(),
+                    required_limits: adapter.limits(),
+                },
+                None,
+            )
+            .await
+            .map_err(|e| JsValue::from_str(&e.to_string()))?;
+        let caps = surface.get_capabilities(&adapter);
+        let config = wgpu::SurfaceConfiguration {
+            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
+            format: caps.formats[0],
+            width: canvas.width(),
+            height: canvas.height(),
+            present_mode: caps.present_modes[0],
+            desired_maximum_frame_latency: 2,
+            alpha_mode: caps.alpha_modes[0],
+            view_formats: vec![],
+        };
+        surface.configure(&device, &config);
+        let aspect = config.width as f32 / config.height as f32;
+
+        let depth_format = wgpu::TextureFormat::Depth32Float;
+        let (depth_texture, depth_view) = graphics::depth::create_depth_texture(
+            &device,
+            depth_format,
+            config.width,
+            config.height,
+        );
+
+        let vertex_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("vertex buffer"),
+            contents: as_bytes(VERTICES),
+            usage: wgpu::BufferUsages::VERTEX | wgpu::BufferUsages::COPY_DST,
+        });
+        let index_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("index buffer"),
+            contents: as_bytes(INDICES),
+            usage: wgpu::BufferUsages::INDEX | wgpu::BufferUsages::COPY_DST,
+        });
+
+        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            label: Some("bind group layout"),
+            entries: &[wgpu::BindGroupLayoutEntry {
+                binding: 0,
+                visibility: wgpu::ShaderStages::VERTEX,
+                ty: wgpu::BindingType::Buffer {
+                    ty: wgpu::BufferBindingType::Uniform,
+                    has_dynamic_offset: false,
+                    min_binding_size: None,
+                },
+                count: None,
+            }],
+        });
+
+        let pipeline = graphics::pipeline::create_pipeline(&device, &config, &bind_group_layout);
+        let uniform = Uniforms {
+            mvp: [
+                [1.0, 0.0, 0.0, 0.0],
+                [0.0, 1.0, 0.0, 0.0],
+                [0.0, 0.0, 1.0, 0.0],
+                [0.0, 0.0, 0.0, 1.0],
+            ],
+        };
+        let uniform_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("uniform buffer"),
+            contents: as_bytes(&[uniform]),
+            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
+        });
+        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+            layout: &bind_group_layout,
+            entries: &[wgpu::BindGroupEntry {
+                binding: 0,
+                resource: uniform_buffer.as_entire_binding(),
+            }],
+            label: Some("bind group"),
+        });
+
+        Ok(Self {
+            surface,
+            device,
+            queue,
+            pipeline,
+            vertex_buffer,
+            index_buffer,
+            uniform_buffer,
+            bind_group,
+            depth_texture,
+            depth_view,
+            depth_format,
+            aspect,
+        })
+    }
+
+    fn update(&mut self, angle: f32, cam_matrix: Mat4) {
+        let model = Mat4::from_rotation_z(angle);
+        let mvp = (cam_matrix * model);
+        let uniform = Uniforms {
+            mvp: mvp.to_cols_array_2d(),
+        };
+        self.queue
+            .write_buffer(&self.uniform_buffer, 0, as_bytes(&[uniform]));
+    }
+
+    fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
+        let frame = self.surface.get_current_texture()?;
+        let view = frame
+            .texture
+            .create_view(&wgpu::TextureViewDescriptor::default());
+        let mut encoder = self
+            .device
+            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
+                label: Some("encoder"),
+            });
+        {
+            let mut rp = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+                label: Some("render"),
+                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
+                    view: &view,
+                    resolve_target: None,
+                    ops: wgpu::Operations {
+                        load: wgpu::LoadOp::Clear(wgpu::Color { r: 0.1, g: 0.1, b: 0.3, a: 1.0 }),
+                        store: wgpu::StoreOp::Store,
+                    },
+                })],
+                depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
+                    view: &self.depth_view,
+                    depth_ops: Some(wgpu::Operations {
+                        load: wgpu::LoadOp::Clear(1.0),
+                        store: wgpu::StoreOp::Store,
+                    }),
+                    stencil_ops: None,
+                }),
+                occlusion_query_set: None,
+                timestamp_writes: None,
+            });
+            rp.set_pipeline(&self.pipeline);
+            rp.set_bind_group(0, &self.bind_group, &[]);
+            rp.set_vertex_buffer(0, self.vertex_buffer.slice(..));
+            rp.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+            rp.draw_indexed(0..INDICES.len() as u32, 0, 0..1);
+        }
+        self.queue.submit(Some(encoder.finish()));
+        frame.present();
+        Ok(())
+    }
+}
+
+#[cfg(target_arch = "wasm32")]
+pub async fn run() -> Result<(), JsValue> {
+    console_error_panic_hook::set_once();
+    web_sys::console::log_1(&"Test 321".into());
+    let window = web_sys::window().unwrap();
+    let document = window.document().unwrap();
+    let canvas = document
+        .get_element_by_id("gpu-canvas")
+        .unwrap()
+        .dyn_into::<web_sys::HtmlCanvasElement>()?;
+
+    let state = Rc::new(RefCell::new(State::new(&canvas).await?));
+    let performance = window.performance().unwrap();
+    let aspect = state.borrow().aspect;
+    let camera = Rc::new(RefCell::new(Camera::new(aspect)));
+
+    crate::input::controls::attach(&window, camera.clone())?;
+
+    let start_time = performance.now();
+    let prev_time = Rc::new(RefCell::new(start_time));
+    let f: Rc<RefCell<Option<Closure<dyn FnMut()>>>> = Rc::new(RefCell::new(None));
+    let g = f.clone();
+    let window_c = window.clone();
+    let perf_c = performance.clone();
+    let camera_c = camera.clone();
+    let prev_time_c = prev_time.clone();
+
+    *g.borrow_mut() = Some(Closure::wrap(Box::new(move || {
+        let now = perf_c.now();
+        let dt = (now - *prev_time_c.borrow()) as f32 / 1000.0;
+        *prev_time_c.borrow_mut() = now;
+        let elapsed = (now - start_time) as f32 / 1000.0;
+        let angle = elapsed / 5.0 * (2.0 * std::f32::consts::PI);
+        {
+            let mut cam = camera_c.borrow_mut();
+            cam.update(dt);
+            let cam_matrix = cam.matrix();
+            let mut st = state.borrow_mut();
+            st.update(angle, cam_matrix);
+            if st.render().is_err() {
+                return;
+            }
+        }
+        window_c
+            .request_animation_frame(f.borrow().as_ref().unwrap().as_ref().unchecked_ref())
+            .unwrap();
+    }) as Box<dyn FnMut()>));
+
+    window.request_animation_frame(g.borrow().as_ref().unwrap().as_ref().unchecked_ref())?;
+    Ok(())
+}
