+++ src/graphics/vertex.rs
@@ -1,0 +1,80 @@
+use wgpu;
+#[repr(C)]
+#[derive(Clone, Copy)]
+pub struct Vertex {
+    pub position: [f32; 3],
+    pub color: [f32; 3],
+}
+
+impl Vertex {
+    pub fn desc<'a>() -> wgpu::VertexBufferLayout<'a> {
+        use std::mem;
+        wgpu::VertexBufferLayout {
+            array_stride: mem::size_of::<Vertex>() as wgpu::BufferAddress,
+            step_mode: wgpu::VertexStepMode::Vertex,
+            attributes: &[
+                wgpu::VertexAttribute {
+                    offset: 0,
+                    shader_location: 0,
+                    format: wgpu::VertexFormat::Float32x3,
+                },
+                wgpu::VertexAttribute {
+                    offset: mem::size_of::<[f32; 3]>() as wgpu::BufferAddress,
+                    shader_location: 1,
+                    format: wgpu::VertexFormat::Float32x3,
+                },
+            ],
+        }
+    }
+}
+
+pub const VERTICES: &[Vertex] = &[
+    // front - red
+    Vertex { position: [-0.5, -0.5, 0.5], color: [1.0, 0.0, 0.0] },
+    Vertex { position: [0.5, -0.5, 0.5], color: [1.0, 0.0, 0.0] },
+    Vertex { position: [0.5, 0.5, 0.5], color: [1.0, 0.0, 0.0] },
+    Vertex { position: [-0.5, 0.5, 0.5], color: [1.0, 0.0, 0.0] },
+    // back - green
+    Vertex { position: [0.5, -0.5, -0.5], color: [0.0, 1.0, 0.0] },
+    Vertex { position: [-0.5, -0.5, -0.5], color: [0.0, 1.0, 0.0] },
+    Vertex { position: [-0.5, 0.5, -0.5], color: [0.0, 1.0, 0.0] },
+    Vertex { position: [0.5, 0.5, -0.5], color: [0.0, 1.0, 0.0] },
+    // left - blue
+    Vertex { position: [-0.5, -0.5, -0.5], color: [0.0, 0.0, 1.0] },
+    Vertex { position: [-0.5, -0.5, 0.5], color: [0.0, 0.0, 1.0] },
+    Vertex { position: [-0.5, 0.5, 0.5], color: [0.0, 0.0, 1.0] },
+    Vertex { position: [-0.5, 0.5, -0.5], color: [0.0, 0.0, 1.0] },
+    // right - yellow
+    Vertex { position: [0.5, -0.5, 0.5], color: [1.0, 1.0, 0.0] },
+    Vertex { position: [0.5, -0.5, -0.5], color: [1.0, 1.0, 0.0] },
+    Vertex { position: [0.5, 0.5, -0.5], color: [1.0, 1.0, 0.0] },
+    Vertex { position: [0.5, 0.5, 0.5], color: [1.0, 1.0, 0.0] },
+    // top - cyan
+    Vertex { position: [-0.5, 0.5, 0.5], color: [0.0, 1.0, 1.0] },
+    Vertex { position: [0.5, 0.5, 0.5], color: [0.0, 1.0, 1.0] },
+    Vertex { position: [0.5, 0.5, -0.5], color: [0.0, 1.0, 1.0] },
+    Vertex { position: [-0.5, 0.5, -0.5], color: [0.0, 1.0, 1.0] },
+    // bottom - magenta
+    Vertex { position: [-0.5, -0.5, -0.5], color: [1.0, 0.0, 1.0] },
+    Vertex { position: [0.5, -0.5, -0.5], color: [1.0, 0.0, 1.0] },
+    Vertex { position: [0.5, -0.5, 0.5], color: [1.0, 0.0, 1.0] },
+    Vertex { position: [-0.5, -0.5, 0.5], color: [1.0, 0.0, 1.0] },
+];
+
+pub const INDICES: &[u16] = &[
+    0, 1, 2, 0, 2, 3,
+    4, 5, 6, 4, 6, 7,
+    8, 9, 10, 8, 10, 11,
+    12, 13, 14, 12, 14, 15,
+    16, 17, 18, 16, 18, 19,
+    20, 21, 22, 20, 22, 23,
+];
+
+pub fn as_bytes<T: Copy>(data: &[T]) -> &[u8] {
+    unsafe {
+        std::slice::from_raw_parts(
+            data.as_ptr() as *const u8,
+            data.len() * std::mem::size_of::<T>(),
+        )
+    }
+}
