+++ src/input/mod.rs
@@ -1,0 +1,2 @@
+pub mod camera;
+pub mod controls;
